Refactor Post & User with Resources
using Resource to control the response data.
with php artisan make:resource [model].
that will make a app\Http\Resources Folder for each model.
This resource class receive a input as $request. and return an array. of data  which point to itself.
The repair in Post model to repair the relation between User and Post.